Fixed compiler bug

+++ src/main/java/ldbc/snb/datagen/objects/similarity/InterestsSimilarity.java
@@ -6,7 +6,7 @@ import ldbc.snb.datagen.objects.Person;
  * Created by aprat on 22/01/16.
  */
 public class InterestsSimilarity implements Person.PersonSimilarity {
-    public double Similarity(Person personA, Person personB) {
+    public float Similarity(Person personA, Person personB) {
         return 0;
     }
 }
